🩹 Fix: map함수   #26
