+++ src/ui/shadcn/button.tsx
@@ -20,13 +20,13 @@ const buttonVariants = cva(
         link: 'text-primary underline-offset-4 hover:underline',
         button: 'bg-main1 !text-title-sm text-light-gray',
         label: 'bg-sub1 !text-text-lg !text-light-gray pointer-events-none',
-        logo: 'bg-main1 !text-title-md text-light-gray',
+        logo: 'bg-main1 text-light-gray',
       },
       size: {
         default: 'px-2 py-1',
         label: 'px-6 py-2 rounded-2xl',
         button: 'px-8 py-3 rounded-2xl',
-        logo: 'px-8 py-6 rounded-2xl',
+        logo: 'px-8 py-5 rounded-2xl',
         icon: 'h-10 w-10',
       },
     },
